Circular import because of find_best_running_fitness
